test: GRAFT spam and GRAFT backoff tests
